requirements.txt
pydantic
python-dotenv
langchain>=0.2.0
langgraph>=0.0.40
openai>=1.0.0          # si usas OpenAI además de AWS
boto3>=1.34.0          # SDK de AWS para llamar a Bedrock/Comprehend/etc.
botocore>=1.34.0      # Dependencia de boto3
networkx
IPython